% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_design.R
\name{gen_sys}
\alias{gen_sys}
\title{Two-dimensional systematic sample}
\usage{
gen_sys(dis, ras_dim, start = NULL)
}
\arguments{
\item{dis}{A vector of length two indicating the distance in pixels between
sampling locations. The first element is the distance in y-direction, the
second element the distance in x-direction.}

\item{ras_dim}{Dimensions of the raster image on which to perform sampling.
The first value is the number of rows, the second the number of columns.}

\item{start}{A vector of length two indicating the starting point of the
sampling grid. The first element is the row index, the second the column
index. If \code{NULL}, a random start is generated.}
}
\value{
A data.table object with row and column indices
}
\description{
Two-dimensional systematic sample
}
\details{
Sampling on irregular shapes is not implemented yet. The function
  assumes a recangular shape of the study area.
}
\author{
Sebastian Schnell
}

