% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_boundary}
\alias{load_boundary}
\title{Load country boundaries into a \link[sp]{\code{SpatialPolygonsDataFrame}}
object}
\usage{
load_boundary(x, country_code, adm_level, ...)
}
\arguments{
\item{x}{If missing, \link[raster]{\code{getData}} function is used to load
boundaries from the GADM project (http://gadm.org). Otherwise the name of
the data source (see \link[gdal]{\code{readOGR}}).}

\item{country_code}{Country specified by three letter ISO code. Use function
\link[raster]{\code{ccodes}} to get appropriate codes.}

\item{adm_level}{Level of administrative subdivision 0 = country, 1 = first
level subdivision}

\item{...}{Other arguments to \link[rgdal]{\code{readOGR}}.}
}
\value{
A \link[sp]{\code{SpatialPolygonsDataFrame}} object storing country
  boundaries.
}
\description{
Load country boundaries into a \link[sp]{\code{SpatialPolygonsDataFrame}}
object
}
\examples{
# Load boundary of Germany
ger_bnd <- load_boundary(x = NA, country_code = "DEU", adm_level = 0);
plot(ger_bnd);
}
\author{
Nils Noelke, Sebastian Schnell
}

