% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sim.R
\name{sim_forest_cover}
\alias{sim_forest_cover}
\title{Simulating forest cover maps with defined covariance structure using random
fields}
\usage{
sim_forest_cover(ras_dim, pxl_size, p, cov_model = NULL, alpha = 0.7,
  beta = 0.3, scale = floor(1/3 * sqrt((sum(ras_dim * pxl_size)^2))), ...)
}
\arguments{
\item{ras_dim}{A vector of two values where the first specifies the number of
columns and the second the number of rows of the output raster.}

\item{pxl_size}{Pixel size of the output map.}

\item{p}{Forest cover of the out put map, 0 < p <= 1.}

\item{cov_model}{A covariance model used for the random fields simulation.
Type \code{RFgetModelNames(type="positive definite")} to get a list of
available models. Defaults to \code{NULL} in which case a generalized
cauchy family covariance model is used
(\code{\link[RandomFields]{RMgencauchy}}).}

\item{alpha}{See \code{\link[RandomFields]{RMgencauchy}}}

\item{beta}{See \code{\link[RandomFields]{RMgencauchy}}}

\item{scale}{See \code{\link[RandomFields]{RMgencauchy}}. Default value is
set to on third of the main diagonal taken from \code{ras_dim} and
\code{pxl_size}.}

\item{...}{Other parameters used by \code{\link[RandomFields]{RMgencauchy}}.}
}
\value{
A binary \code{\link[raster]{raster}} map of simulated forest cover.
}
\description{
Simulating forest cover maps with defined covariance structure using random
fields
}
\examples{
# Raster size
ras_dim <- c(1000, 1000);

# Pixel size
pxl_size <- 30;

# Forest cover
p <- 0.7;
# Cauchy covariance model
# 0 < alpha <= 2 - low alpha -> high fragmentation
# beta > 0 - smaller beta -> longer long-range dependence
cov_model <- RMgencauchy(alpha = 1.95,
                         beta = 0.3,
                         scale = 1/3*sqrt((sum(ras_dim*pxl_size)^2)));
forest_ras <- sim_forest_cover(ras_dim, pxl_size, cov_model, p);
plot(forest_ras);
}
\author{
Paul Magdon, Sebastian Schnell
}
