% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_boundary}
\alias{load_boundary}
\title{Load country boundaries into a \code{\link[sp]{SpatialPolygonsDataFrame}}
object}
\usage{
load_boundary(x = NULL, country_code = "world", adm_level = 0, ...)
}
\arguments{
\item{x}{If missing, \code{link[raster]{getData}} function is used to load
boundaries from the GADM project (http://gadm.org). Otherwise the name of
the data source (see \code{\link[gdal]{readOGR}}).}

\item{country_code}{Country specified by three letter ISO code. Use function
\code{\link[raster]{ccodes}} to get appropriate codes. If multiple
countries are desired, provide a character vector of appropriate ISO codes.}

\item{adm_level}{Level of administrative subdivision 0 = country, 1 = first
level subdivision}

\item{...}{Other arguments to \code{\link[rgdal]{readOGR}}.}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object storing country
  boundaries.
}
\description{
Load country boundaries into a \code{\link[sp]{SpatialPolygonsDataFrame}}
object
}
\examples{
# Load boundary of Germany
ger_bnd <- load_boundary(country_code = "DEU", adm_level = 0);
plot(ger_bnd);

# Load boundary of several countries
bnd <- load_boundary(country_code = c("DEU", "ITA", "FRA"));
}
\author{
Nils Noelke, Sebastian Schnell
}
